Fix bugs and refactor generation utils
